Experiment with potentially useful/necessary elements.
Experiment with input element to get user instructions. Try buttons in and out of forms.Button in form
should be of type 'button' for this application. Use 'preventDefault' in eventListener added to Enter
key press. Use e.target instead of 'this' to identify div as target of mouseover or mouseout. Discovered
'document.Fragment' to use to build grid of divs, and then add fragment to container div, rather than
add each div element individually, and avoid 're-painting' the page over and over.